pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai2-molmo"
dynamic = ["version"]
readme = "README.md"
description = "Multimodal Open Language Model (Molmo)"
authors = [
    { name = "Allen Institute for Artificial Intelligence" }
]
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "absl-py",
    "cached_property",
    "fiddle >= 0.2.5",
    "numpy",
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "omegaconf",
    "rich",
    "boto3",
    "google-cloud-storage",
    "tokenizers",
    "packaging",
    "cached_path>=1.6.4",
    "transformers>=v4.37.1",
    "datasets",
    "huggingface_hub",
    "sentencepiece",
    "requests",
    "einops",
    "einops-exts",
    "tqdm",
    "gcsfs==2023.9.2",
    "accelerate",
    "imageio",
    "imageio[pyav]",
    "decord",
    "moviepy",
    "prior",
    "Pillow",
    "av",
    "torchmetrics",
    "modal",
    "h5py",
    "splitstream",
    "ffmpeg-python>=0.2.0",
    "scipy"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy>=1.0,<1.4",
    "black>=23.1,<24.0",
    "isort>=5.12,<5.13",
    "pytest",
    "pytest-sphinx",
    "twine>=1.11.0",
    "setuptools",
    "wheel",
    "build",
]
train = [
    "wandb",
    "beaker-py>=1.34.1",
    "beaker-gantry",
    "click",
    "torchmetrics",
    "smashed[remote]>=0.21.1",
    "safetensors",
    "scikit-learn",
    "msgspec>=0.14.0",
    "openai",
    "python-Levenshtein",
    "editdistance"
]
serve = [
    "pydantic",
    "uvicorn",
    "gradio",
    "fastapi",
]
all = [
    "ai2-molmo[dev,train,serve]",
]

[project.urls]
Homepage = "https://github.com/allenai/molmo"
Repository = "https://github.com/allenai/molmo"

[tool.setuptools]
include-package-data = true
packages = ["olmo"]
